mt issue ressolved

+++ lib/presentation/features/leads_screen/view/AddNewLeadForm.dart
@@ -1932,9 +1932,7 @@ class _AddNewLeadFormState extends State<AddNewLeadForm> {
                                       _totalMT.text == "") {
                                     _totalBags.clear();
                                   } else {
-                                    _totalBags.text =
-                                        (int.parse(_totalMT.text) * 20)
-                                            .toString();
+                                    _totalBags.text = (double.parse(_totalMT.text) * 20).toInt().toString();
                                   }
                                 });
                               },
