+++ lib/presentation/features/leads_screen/view/ViewLeadScreen.dart
@@ -3133,9 +3133,8 @@ class _ViewLeadScreenState extends State<ViewLeadScreen>
                                       _totalMT.text == "") {
                                     _totalBags.clear();
                                   } else {
-                                    _totalBags.text =
-                                        (int.parse(_totalMT.text) * 20)
-                                            .toString();
+
+                                    _totalBags.text = (double.parse(_totalMT.text) * 20).toInt().toString();
                                   }
                                 });
                               },
