+++ lib/presentation/features/leads_screen/view/ViewOldLeadScreen.dart
@@ -2224,9 +2224,7 @@ class _ViewOldLeadScreeState extends State<ViewOldLeadScree> {
                                         _totalMT.text == "") {
                                       _totalBags.clear();
                                     } else {
-                                      _totalBags.text =
-                                          (int.parse(_totalMT.text) * 20)
-                                              .toString();
+                                      _totalBags.text =(double.parse(_totalMT.text) * 20).toInt().toString();
                                     }
                                   });
                                 },
